+++ src/ServiceProvider.php
@@ -1,14 +1,19 @@
 <?php
+
 namespace DreamFactory\Core\File;
 
 use DreamFactory\Core\Components\ServiceDocBuilder;
 use DreamFactory\Core\Enums\ServiceTypeGroups;
-use DreamFactory\Core\Handlers\Events\ServiceEventHandler;
+use DreamFactory\Core\File\Models\FTPFileConfig;
+use DreamFactory\Core\File\Models\SFTPFileConfig;
+use DreamFactory\Core\File\Models\WebDAVFileConfig;
+use DreamFactory\Core\File\Services\FTPFileService;
+use DreamFactory\Core\File\Services\SFTPFileService;
 use DreamFactory\Core\File\Models\LocalFileConfig;
 use DreamFactory\Core\File\Services\LocalFileService;
+use DreamFactory\Core\File\Services\WebDAVFileService;
 use DreamFactory\Core\Services\ServiceManager;
 use DreamFactory\Core\Services\ServiceType;
-use Event;
 
 class ServiceProvider extends \Illuminate\Support\ServiceProvider
 {
@@ -45,7 +50,7 @@ class ServiceProvider extends \Illuminate\Support\ServiceProvider
         $this->mergeConfigFrom($configPath, 'df');
 
         // Add our service types.
-        $this->app->resolving('df.service', function (ServiceManager $df) {
+        $this->app->resolving('df.service', function (ServiceManager $df){
             $df->addType(
                 new ServiceType([
                     'name'            => 'local_file',
@@ -53,11 +58,59 @@ class ServiceProvider extends \Illuminate\Support\ServiceProvider
                     'description'     => 'File service supporting the local file system.',
                     'group'           => ServiceTypeGroups::FILE,
                     'config_handler'  => LocalFileConfig::class,
-                    'default_api_doc' => function ($service) {
+                    'default_api_doc' => function ($service){
                         return $this->buildServiceDoc($service->id, LocalFileService::getApiDocInfo($service));
                     },
-                    'factory'         => function ($config) {
+                    'factory'         => function ($config){
                         return new LocalFileService($config);
+                    },
+                ])
+            );
+
+            $df->addType(
+                new ServiceType([
+                    'name'            => 'ftp_file',
+                    'label'           => 'FTP File Storage',
+                    'description'     => 'File service supporting the FTP protocol.',
+                    'group'           => ServiceTypeGroups::FILE,
+                    'config_handler'  => FTPFileConfig::class,
+                    'default_api_doc' => function ($service){
+                        return $this->buildServiceDoc($service->id, FTPFileService::getApiDocInfo($service));
+                    },
+                    'factory'         => function ($config){
+                        return new FTPFileService($config);
+                    },
+                ])
+            );
+
+            $df->addType(
+                new ServiceType([
+                    'name'            => 'sftp_file',
+                    'label'           => 'SFTP File Storage',
+                    'description'     => 'File service supporting the SFTP protocol.',
+                    'group'           => ServiceTypeGroups::FILE,
+                    'config_handler'  => SFTPFileConfig::class,
+                    'default_api_doc' => function ($service){
+                        return $this->buildServiceDoc($service->id, SFTPFileService::getApiDocInfo($service));
+                    },
+                    'factory'         => function ($config){
+                        return new SFTPFileService($config);
+                    },
+                ])
+            );
+
+            $df->addType(
+                new ServiceType([
+                    'name'            => 'webdav_file',
+                    'label'           => 'WebDAV File Storage',
+                    'description'     => 'File service supporting WebDAV.',
+                    'group'           => ServiceTypeGroups::FILE,
+                    'config_handler'  => WebDAVFileConfig::class,
+                    'default_api_doc' => function ($service){
+                        return $this->buildServiceDoc($service->id, WebDAVFileService::getApiDocInfo($service));
+                    },
+                    'factory'         => function ($config){
+                        return new WebDAVFileService($config);
                     },
                 ])
             );