+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.3.0] - 2017-07-27
+### Fixed
+- DF-1144 Made DELETE behavior consistent across local and all remote file services
+- Fixed file upload issue for FTP, SFTP, and WebDAV
+- Do not calculate path on container level files
+### Added
+- DF-580 Added support for FTP, SFTP, FTPS
+- DF-276 Added support for WebDAV
+
 ## [0.2.1] - 2017-06-21
 ### Changed
 - Local file service config now accepts paths relative to the storage folder as well as full system paths.
@@ -22,6 +31,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 First official release working with the new [dreamfactory](https://github.com/dreamfactorysoftware/dreamfactory) project.
 
-[Unreleased]: https://github.com/dreamfactorysoftware/df-file/compare/0.2.1...HEAD
+[Unreleased]: https://github.com/dreamfactorysoftware/df-file/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/dreamfactorysoftware/df-file/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/dreamfactorysoftware/df-file/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/dreamfactorysoftware/df-file/compare/0.1.0...0.2.0